+++ include/bio/alphabet/composite/all.hpp
@@ -12,9 +12,9 @@
  */
 #pragma once
 
-#include <bio/alphabet/composite/alphabet_tuple_base.hpp>
-#include <bio/alphabet/composite/alphabet_variant.hpp>
 #include <bio/alphabet/composite/semialphabet_any.hpp>
+#include <bio/alphabet/composite/tuple_base.hpp>
+#include <bio/alphabet/composite/variant.hpp>
 
 /*!\defgroup alphabet_composite Composite
  * \brief Provides templates for combining existing alphabets into new alphabet types.
@@ -25,13 +25,13 @@
  * ### Introduction
  *
  * This module provides various class templates that allow you to combine existing alphabets into new ones. For example,
- * you can add new characters to existing alphabets by using bio::alphabet::alphabet_variant or combine alphabets with quality
- * information by using bio::alphabet::alphabet_tuple_base.
+ * you can add new characters to existing alphabets by using bio::alphabet::variant or combine alphabets with quality
+ * information by using bio::alphabet::tuple_base.
  *
  * We have currently three major composite alphabets:
- * * bio::alphabet::alphabet_tuple_base which can be used to create a std::tuple like object that still models
+ * * bio::alphabet::tuple_base which can be used to create a std::tuple like object that still models
  *   bio::alphabet::alphabet.
- * * bio::alphabet::alphabet_variant which roughly corresponds to the Union of the given types. It behaves similar to
+ * * bio::alphabet::variant which roughly corresponds to the Union of the given types. It behaves similar to
  *   std::variant, but also models bio::alphabet::alphabet.
  * * bio::alphabet::semialphabet_any which type erases other alphabets of the same size and allows again transformation to
  *   alphabets of the same size by copying the rank.