+++ include/bio/ranges/views/repeat.hpp
@@ -236,20 +236,12 @@ public:
      */
     //!\brief Inherit the equality comparison (same type) from base type.
     using base_t::operator==;
-    //!\brief Inherit the inequality comparison (same type) from base type.
-    using base_t::operator!=;
 
     //!\brief Equality comparison to the sentinel always returns false on an infinite view.
     constexpr bool operator==(std::default_sentinel_t const &) const noexcept { return false; }
 
-    //!\brief Inequality comparison to the sentinel always returns true on an infinite view.
-    constexpr bool operator!=(std::default_sentinel_t const &) const noexcept { return true; }
-
     //!\brief Equality comparison to the sentinel always returns false on an infinite view.
     friend constexpr bool operator==(std::default_sentinel_t const &, basic_iterator const &) noexcept { return false; }
-
-    //!\brief Inequality comparison to the sentinel always returns true on an infinite view.
-    friend constexpr bool operator!=(std::default_sentinel_t const &, basic_iterator const &) noexcept { return true; }
     //!\}
 };