+++ test/performance/ranges/container_push_back_benchmark.cpp
@@ -49,7 +49,7 @@ BENCHMARK_TEMPLATE(push_back, std::vector, bio::alphabet::dna4);
 BENCHMARK_TEMPLATE(push_back, std::vector, bio::alphabet::gapped<bio::alphabet::dna4>);
 BENCHMARK_TEMPLATE(push_back, std::vector, bio::alphabet::dna15);
 BENCHMARK_TEMPLATE(push_back, std::vector, bio::alphabet::aa27);
-BENCHMARK_TEMPLATE(push_back, std::vector, bio::alphabet::alphabet_variant<char, bio::alphabet::dna4>);
+BENCHMARK_TEMPLATE(push_back, std::vector, bio::alphabet::variant<char, bio::alphabet::dna4>);
 
 BENCHMARK_TEMPLATE(push_back, std::deque, char);
 BENCHMARK_TEMPLATE(push_back, std::deque, uint8_t);
@@ -61,7 +61,7 @@ BENCHMARK_TEMPLATE(push_back, std::deque, bio::alphabet::dna4);
 BENCHMARK_TEMPLATE(push_back, std::deque, bio::alphabet::gapped<bio::alphabet::dna4>);
 BENCHMARK_TEMPLATE(push_back, std::deque, bio::alphabet::dna15);
 BENCHMARK_TEMPLATE(push_back, std::deque, bio::alphabet::aa27);
-BENCHMARK_TEMPLATE(push_back, std::deque, bio::alphabet::alphabet_variant<char, bio::alphabet::dna4>);
+BENCHMARK_TEMPLATE(push_back, std::deque, bio::alphabet::variant<char, bio::alphabet::dna4>);
 
 BENCHMARK_TEMPLATE(push_back, std::list, char);
 BENCHMARK_TEMPLATE(push_back, std::list, uint8_t);
@@ -73,7 +73,7 @@ BENCHMARK_TEMPLATE(push_back, std::list, bio::alphabet::dna4);
 BENCHMARK_TEMPLATE(push_back, std::list, bio::alphabet::gapped<bio::alphabet::dna4>);
 BENCHMARK_TEMPLATE(push_back, std::list, bio::alphabet::dna15);
 BENCHMARK_TEMPLATE(push_back, std::list, bio::alphabet::aa27);
-BENCHMARK_TEMPLATE(push_back, std::list, bio::alphabet::alphabet_variant<char, bio::alphabet::dna4>);
+BENCHMARK_TEMPLATE(push_back, std::list, bio::alphabet::variant<char, bio::alphabet::dna4>);
 
 #if __has_include(<sdsl/int_vector.hpp>)
 #include <sdsl/int_vector.hpp>
@@ -92,7 +92,7 @@ BENCHMARK_TEMPLATE(push_back, bio::ranges::bitcompressed_vector, bio::alphabet::
 BENCHMARK_TEMPLATE(push_back, bio::ranges::bitcompressed_vector, bio::alphabet::gapped<bio::alphabet::dna4>);
 BENCHMARK_TEMPLATE(push_back, bio::ranges::bitcompressed_vector, bio::alphabet::dna15);
 BENCHMARK_TEMPLATE(push_back, bio::ranges::bitcompressed_vector, bio::alphabet::aa27);
-BENCHMARK_TEMPLATE(push_back, bio::ranges::bitcompressed_vector, bio::alphabet::alphabet_variant<char, bio::alphabet::dna4>);
+BENCHMARK_TEMPLATE(push_back, bio::ranges::bitcompressed_vector, bio::alphabet::variant<char, bio::alphabet::dna4>);
 
 BENCHMARK_TEMPLATE(push_back, small_vec, char);
 BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::gap);
@@ -100,7 +100,7 @@ BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::dna4);
 BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::gapped<bio::alphabet::dna4>);
 BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::dna15);
 BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::aa27);
-BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::alphabet_variant<char, bio::alphabet::dna4>);
+BENCHMARK_TEMPLATE(push_back, small_vec, bio::alphabet::variant<char, bio::alphabet::dna4>);
 
 // ============================================================================
 //  run